doc: moved commands to another folder
