+++ Crescendo2024/src/main/java/frc/robot/commands/teleopCommands/arm/Level0Setpoint.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.teleopCommands;
+package frc.robot.commands.teleopCommands.arm;
 
 import frc.robot.FireBirdsUtils;
 import frc.robot.Constants.ArmConstants;
